feat/upgrade smart contract (#67)
* feat: upgrade smart contract version, role and types

* refactor: keep droplet generated types (new version does not include it)

* fix: adjust new droplet types imports
