+++ src/components/create-pool/create-pool.tsx
@@ -18,7 +18,8 @@ import { useWaitForTransactionReceipt, useWriteContract } from 'wagmi'
 import { createPoolAction, updatePoolStatus } from './action'
 import { toast } from 'sonner'
 import { getAbiItem, parseEther } from 'viem'
-import { dropletAddress, poolAbi, poolAddress } from '@/types/contracts'
+import { poolAbi, poolAddress } from '@/types/contracts'
+import { dropletAddress } from '@/types/droplet'
 import { wagmi } from '@/providers/configs'
 import { useRouter } from 'next/navigation'
