+++ src/components/pool-details/pool-details.tsx
@@ -9,7 +9,8 @@ import { formatEventDateTime } from '@/lib/utils/date-time'
 import { cn } from '@/lib/utils/tailwind'
 import { useBottomBarStore } from '@/providers/bottom-bar.provider'
 import { wagmi } from '@/providers/configs'
-import { dropletAbi, dropletAddress, poolAbi, poolAddress } from '@/types/contracts'
+import { dropletAbi, dropletAddress } from '@/types/droplet'
+import { poolAbi, poolAddress } from '@/types/contracts'
 import { useWallets } from '@privy-io/react-auth'
 import { useQueryClient } from '@tanstack/react-query'
 import { ChevronRight } from 'lucide-react'
