+++ src/types/contracts.ts
@@ -1,367 +1,9 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Droplet
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0xfD2Ec58cE4c87b253567Ff98ce2778de6AF0101b)
- */
-export const dropletAbi = [
-    { type: 'constructor', inputs: [], stateMutability: 'nonpayable' },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'owner', internalType: 'address', type: 'address' },
-            { name: 'spender', internalType: 'address', type: 'address' },
-        ],
-        name: 'allowance',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'spender', internalType: 'address', type: 'address' },
-            { name: 'value', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'approve',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [{ name: 'account', internalType: 'address', type: 'address' }],
-        name: 'balanceOf',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'decimals',
-        outputs: [{ name: '', internalType: 'uint8', type: 'uint8' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'to', internalType: 'address', type: 'address' },
-            { name: 'amount', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'mint',
-        outputs: [],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'name',
-        outputs: [{ name: '', internalType: 'string', type: 'string' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'symbol',
-        outputs: [{ name: '', internalType: 'string', type: 'string' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'totalSupply',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'to', internalType: 'address', type: 'address' },
-            { name: 'value', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'transfer',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'from', internalType: 'address', type: 'address' },
-            { name: 'to', internalType: 'address', type: 'address' },
-            { name: 'value', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'transferFrom',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'event',
-        anonymous: false,
-        inputs: [
-            {
-                name: 'owner',
-                internalType: 'address',
-                type: 'address',
-                indexed: true,
-            },
-            {
-                name: 'spender',
-                internalType: 'address',
-                type: 'address',
-                indexed: true,
-            },
-            {
-                name: 'value',
-                internalType: 'uint256',
-                type: 'uint256',
-                indexed: false,
-            },
-        ],
-        name: 'Approval',
-    },
-    {
-        type: 'event',
-        anonymous: false,
-        inputs: [
-            { name: 'from', internalType: 'address', type: 'address', indexed: true },
-            { name: 'to', internalType: 'address', type: 'address', indexed: true },
-            {
-                name: 'value',
-                internalType: 'uint256',
-                type: 'uint256',
-                indexed: false,
-            },
-        ],
-        name: 'Transfer',
-    },
-    {
-        type: 'error',
-        inputs: [
-            { name: 'spender', internalType: 'address', type: 'address' },
-            { name: 'allowance', internalType: 'uint256', type: 'uint256' },
-            { name: 'needed', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'ERC20InsufficientAllowance',
-    },
-    {
-        type: 'error',
-        inputs: [
-            { name: 'sender', internalType: 'address', type: 'address' },
-            { name: 'balance', internalType: 'uint256', type: 'uint256' },
-            { name: 'needed', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'ERC20InsufficientBalance',
-    },
-    {
-        type: 'error',
-        inputs: [{ name: 'approver', internalType: 'address', type: 'address' }],
-        name: 'ERC20InvalidApprover',
-    },
-    {
-        type: 'error',
-        inputs: [{ name: 'receiver', internalType: 'address', type: 'address' }],
-        name: 'ERC20InvalidReceiver',
-    },
-    {
-        type: 'error',
-        inputs: [{ name: 'sender', internalType: 'address', type: 'address' }],
-        name: 'ERC20InvalidSender',
-    },
-    {
-        type: 'error',
-        inputs: [{ name: 'spender', internalType: 'address', type: 'address' }],
-        name: 'ERC20InvalidSpender',
-    },
-] as const
-
-/**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0xfD2Ec58cE4c87b253567Ff98ce2778de6AF0101b)
- */
-export const dropletAddress = {
-    8453: '0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512',
-    84532: '0xfD2Ec58cE4c87b253567Ff98ce2778de6AF0101b',
-} as const
-
-/**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0xfD2Ec58cE4c87b253567Ff98ce2778de6AF0101b)
- */
-export const dropletConfig = {
-    address: dropletAddress,
-    abi: dropletAbi,
-} as const
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// MockERC20
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-export const mockErc20Abi = [
-    {
-        type: 'function',
-        inputs: [],
-        name: 'DOMAIN_SEPARATOR',
-        outputs: [{ name: '', internalType: 'bytes32', type: 'bytes32' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'owner', internalType: 'address', type: 'address' },
-            { name: 'spender', internalType: 'address', type: 'address' },
-        ],
-        name: 'allowance',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'spender', internalType: 'address', type: 'address' },
-            { name: 'amount', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'approve',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [{ name: 'owner', internalType: 'address', type: 'address' }],
-        name: 'balanceOf',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'decimals',
-        outputs: [{ name: '', internalType: 'uint8', type: 'uint8' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'name_', internalType: 'string', type: 'string' },
-            { name: 'symbol_', internalType: 'string', type: 'string' },
-            { name: 'decimals_', internalType: 'uint8', type: 'uint8' },
-        ],
-        name: 'initialize',
-        outputs: [],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'name',
-        outputs: [{ name: '', internalType: 'string', type: 'string' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [{ name: '', internalType: 'address', type: 'address' }],
-        name: 'nonces',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'owner', internalType: 'address', type: 'address' },
-            { name: 'spender', internalType: 'address', type: 'address' },
-            { name: 'value', internalType: 'uint256', type: 'uint256' },
-            { name: 'deadline', internalType: 'uint256', type: 'uint256' },
-            { name: 'v', internalType: 'uint8', type: 'uint8' },
-            { name: 'r', internalType: 'bytes32', type: 'bytes32' },
-            { name: 's', internalType: 'bytes32', type: 'bytes32' },
-        ],
-        name: 'permit',
-        outputs: [],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'symbol',
-        outputs: [{ name: '', internalType: 'string', type: 'string' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [],
-        name: 'totalSupply',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'to', internalType: 'address', type: 'address' },
-            { name: 'amount', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'transfer',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'function',
-        inputs: [
-            { name: 'from', internalType: 'address', type: 'address' },
-            { name: 'to', internalType: 'address', type: 'address' },
-            { name: 'amount', internalType: 'uint256', type: 'uint256' },
-        ],
-        name: 'transferFrom',
-        outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-        stateMutability: 'nonpayable',
-    },
-    {
-        type: 'event',
-        anonymous: false,
-        inputs: [
-            {
-                name: 'owner',
-                internalType: 'address',
-                type: 'address',
-                indexed: true,
-            },
-            {
-                name: 'spender',
-                internalType: 'address',
-                type: 'address',
-                indexed: true,
-            },
-            {
-                name: 'value',
-                internalType: 'uint256',
-                type: 'uint256',
-                indexed: false,
-            },
-        ],
-        name: 'Approval',
-    },
-    {
-        type: 'event',
-        anonymous: false,
-        inputs: [
-            { name: 'from', internalType: 'address', type: 'address', indexed: true },
-            { name: 'to', internalType: 'address', type: 'address', indexed: true },
-            {
-                name: 'value',
-                internalType: 'uint256',
-                type: 'uint256',
-                indexed: false,
-            },
-        ],
-        name: 'Transfer',
-    },
-] as const
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Pool
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x44432A98ea8dA37F844B89A324204ee6642b785A)
+ * [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x5C22662210E48D0f5614cACA6f7a6a938716Ea26)
  */
 export const poolAbi = [
     { type: 'constructor', inputs: [], stateMutability: 'nonpayable' },
@@ -375,7 +17,14 @@ export const poolAbi = [
     {
         type: 'function',
         inputs: [],
-        name: 'WHITELISTED',
+        name: 'WHITELISTED_HOST',
+        outputs: [{ name: '', internalType: 'bytes32', type: 'bytes32' }],
+        stateMutability: 'view',
+    },
+    {
+        type: 'function',
+        inputs: [],
+        name: 'WHITELISTED_SPONSOR',
         outputs: [{ name: '', internalType: 'bytes32', type: 'bytes32' }],
         stateMutability: 'view',
     },
@@ -577,7 +226,7 @@ export const poolAbi = [
                     { name: 'feesAccumulated', internalType: 'uint256', type: 'uint256' },
                     { name: 'feesCollected', internalType: 'uint256', type: 'uint256' },
                     { name: 'balance', internalType: 'uint256', type: 'uint256' },
-                    { name: 'extraBalance', internalType: 'uint256', type: 'uint256' },
+                    { name: 'sponsored', internalType: 'uint256', type: 'uint256' },
                 ],
             },
             {
@@ -599,13 +248,6 @@ export const poolAbi = [
             { name: '', internalType: 'uint256[]', type: 'uint256[]' },
             { name: '', internalType: 'bool[]', type: 'bool[]' },
         ],
-        stateMutability: 'view',
-    },
-    {
-        type: 'function',
-        inputs: [{ name: 'poolId', internalType: 'uint256', type: 'uint256' }],
-        name: 'getExtraBalance',
-        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
         stateMutability: 'view',
     },
     {
@@ -733,6 +375,40 @@ export const poolAbi = [
         inputs: [{ name: 'role', internalType: 'bytes32', type: 'bytes32' }],
         name: 'getRoleAdmin',
         outputs: [{ name: '', internalType: 'bytes32', type: 'bytes32' }],
+        stateMutability: 'view',
+    },
+    {
+        type: 'function',
+        inputs: [
+            { name: 'poolId', internalType: 'uint256', type: 'uint256' },
+            { name: '_sponsor', internalType: 'address', type: 'address' },
+        ],
+        name: 'getSponsorDetail',
+        outputs: [
+            {
+                name: '',
+                internalType: 'struct IPool.SponsorDetail',
+                type: 'tuple',
+                components: [
+                    { name: 'name', internalType: 'string', type: 'string' },
+                    { name: 'amount', internalType: 'uint256', type: 'uint256' },
+                ],
+            },
+        ],
+        stateMutability: 'view',
+    },
+    {
+        type: 'function',
+        inputs: [{ name: 'poolId', internalType: 'uint256', type: 'uint256' }],
+        name: 'getSponsors',
+        outputs: [{ name: '', internalType: 'address[]', type: 'address[]' }],
+        stateMutability: 'view',
+    },
+    {
+        type: 'function',
+        inputs: [{ name: 'poolId', internalType: 'uint256', type: 'uint256' }],
+        name: 'getSponsorshipAmount',
+        outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
         stateMutability: 'view',
     },
     {
@@ -928,7 +604,7 @@ export const poolAbi = [
             { name: 'feesAccumulated', internalType: 'uint256', type: 'uint256' },
             { name: 'feesCollected', internalType: 'uint256', type: 'uint256' },
             { name: 'balance', internalType: 'uint256', type: 'uint256' },
-            { name: 'extraBalance', internalType: 'uint256', type: 'uint256' },
+            { name: 'sponsored', internalType: 'uint256', type: 'uint256' },
         ],
         stateMutability: 'view',
     },
@@ -1035,6 +711,40 @@ export const poolAbi = [
         name: 'setWinners',
         outputs: [],
         stateMutability: 'nonpayable',
+    },
+    {
+        type: 'function',
+        inputs: [
+            { name: 'name', internalType: 'string', type: 'string' },
+            { name: 'poolId', internalType: 'uint256', type: 'uint256' },
+            { name: 'amount', internalType: 'uint256', type: 'uint256' },
+        ],
+        name: 'sponsor',
+        outputs: [],
+        stateMutability: 'nonpayable',
+    },
+    {
+        type: 'function',
+        inputs: [
+            { name: '', internalType: 'address', type: 'address' },
+            { name: 'poolId', internalType: 'uint256', type: 'uint256' },
+        ],
+        name: 'sponsorDetail',
+        outputs: [
+            { name: 'name', internalType: 'string', type: 'string' },
+            { name: 'amount', internalType: 'uint256', type: 'uint256' },
+        ],
+        stateMutability: 'view',
+    },
+    {
+        type: 'function',
+        inputs: [
+            { name: '', internalType: 'uint256', type: 'uint256' },
+            { name: '', internalType: 'uint256', type: 'uint256' },
+        ],
+        name: 'sponsors',
+        outputs: [{ name: '', internalType: 'address', type: 'address' }],
+        stateMutability: 'view',
     },
     {
         type: 'function',
@@ -1161,12 +871,6 @@ export const poolAbi = [
                 name: 'fees',
                 internalType: 'uint256',
                 type: 'uint256',
-                indexed: false,
-            },
-            {
-                name: 'chargedAll',
-                internalType: 'bool',
-                type: 'bool',
                 indexed: false,
             },
         ],
@@ -1299,6 +1003,31 @@ export const poolAbi = [
             },
         ],
         name: 'Paused',
+    },
+    {
+        type: 'event',
+        anonymous: false,
+        inputs: [
+            {
+                name: 'poolId',
+                internalType: 'uint256',
+                type: 'uint256',
+                indexed: false,
+            },
+            {
+                name: 'balanceBefore',
+                internalType: 'uint256',
+                type: 'uint256',
+                indexed: false,
+            },
+            {
+                name: 'balanceAfter',
+                internalType: 'uint256',
+                type: 'uint256',
+                indexed: false,
+            },
+        ],
+        name: 'PoolBalanceUpdated',
     },
     {
         type: 'event',
@@ -1524,6 +1253,31 @@ export const poolAbi = [
         anonymous: false,
         inputs: [
             {
+                name: 'poolId',
+                internalType: 'uint256',
+                type: 'uint256',
+                indexed: false,
+            },
+            {
+                name: 'sponsor',
+                internalType: 'address',
+                type: 'address',
+                indexed: true,
+            },
+            {
+                name: 'amount',
+                internalType: 'uint256',
+                type: 'uint256',
+                indexed: false,
+            },
+        ],
+        name: 'SponsorshipAdded',
+    },
+    {
+        type: 'event',
+        anonymous: false,
+        inputs: [
+            {
                 name: 'account',
                 internalType: 'address',
                 type: 'address',
@@ -1645,16 +1399,13 @@ export const poolAbi = [
 ] as const
 
 /**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x44432A98ea8dA37F844B89A324204ee6642b785A)
+ * [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x5C22662210E48D0f5614cACA6f7a6a938716Ea26)
  */
 export const poolAddress = {
-    8453: '0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0',
-    84532: '0x44432A98ea8dA37F844B89A324204ee6642b785A',
+    84532: '0x5C22662210E48D0f5614cACA6f7a6a938716Ea26',
 } as const
 
 /**
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0)
- * - [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x44432A98ea8dA37F844B89A324204ee6642b785A)
+ * [__View Contract on Base Sepolia Basescan__](https://sepolia.basescan.org/address/0x5C22662210E48D0f5614cACA6f7a6a938716Ea26)
  */
 export const poolConfig = { address: poolAddress, abi: poolAbi } as const
